Egon2/pyproject.toml
[project]
name = "egon2"
version = "0.1.0"
description = "Egon der 2. — persönlicher KI-Assistent"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.20",
    "httpx>=0.27,<0.28",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "apscheduler>=3.10,<4",
    "sqlalchemy>=2.0",
    "matrix-nio[e2e]>=0.25",
    "python-telegram-bot==21.*",
    "asyncssh>=2.18,<3",
    "structlog>=25.0",
    "aiolimiter>=1.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.25",
    "respx>=0.22",
    "asgi-lifespan>=2",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["egon2"]
